+++ src/main/scala/edu/ie3/osmogrid/lv/region_coordinator/OsmoGridModelPartitioner.scala
@@ -44,42 +44,33 @@ object OsmoGridModelPartitioner extends LazyLogging {
       osmoGridModel: LvOsmoGridModel,
       areas: ParMap[AreaKey, Seq[Polygon]]
   ): ParMap[AreaKey, LvOsmoGridModel] = {
-    val buildings = assign(
-      osmoGridModel.buildings,
-      areas,
-      AssignByMax
-    )
-    val highways = assign(
-      osmoGridModel.highways,
-      areas,
-      AssignToAll
-    )
-    val landuses = assign(
-      osmoGridModel.landuses,
-      areas,
-      AssignToAll
-    )
-    val boundaries = assign(
-      osmoGridModel.boundaries,
-      areas,
-      AssignByMax
-    )
-    val existingSubstations = assign(
-      osmoGridModel.existingSubstations,
-      areas,
-      AssignByMax
-    )
 
-    areas.keys.map { areaId =>
-      areaId ->
-        LvOsmoGridModel(
-          buildings.getOrElse(areaId, ParSeq.empty),
-          highways.getOrElse(areaId, ParSeq.empty),
-          landuses.getOrElse(areaId, ParSeq.empty),
-          boundaries.getOrElse(areaId, ParSeq.empty),
-          existingSubstations.getOrElse(areaId, ParSeq.empty),
-          osmoGridModel.filter
+    val buildings = assign(osmoGridModel.buildings, areas, AssignByMax)
+    val existingSubstations =
+      assign(osmoGridModel.existingSubstations, areas, AssignByMax)
+    val highways = assign(osmoGridModel.highways, areas, AssignToAll)
+    val landuses = assign(osmoGridModel.landuses, areas, AssignToAll)
+    val boundaries = assign(osmoGridModel.boundaries, areas, AssignByMax)
+
+    areas.keys.flatMap { areaId =>
+      val assignedBuildings = buildings.getOrElse(areaId, ParSeq.empty)
+      val assignedSubstations =
+        existingSubstations.getOrElse(areaId, ParSeq.empty)
+
+      if (assignedBuildings.nonEmpty || assignedSubstations.nonEmpty) {
+        Some(
+          areaId -> LvOsmoGridModel(
+            assignedBuildings,
+            highways.getOrElse(areaId, ParSeq.empty),
+            landuses.getOrElse(areaId, ParSeq.empty),
+            boundaries.getOrElse(areaId, ParSeq.empty),
+            assignedSubstations,
+            osmoGridModel.filter
+          )
         )
+      } else {
+        None
+      }
     }.toMap
   }